The idea of this branch is to loop over xml files in a folder. The item reader reads one at a time, and processes it, and depending on the chunk size, will write out the processed files.
As you can see, I have used standard SB MultiItemReader for the reading.
But, the AdaptiveQMMMProcessor had to be modified, and the JobBuilder
had to be modified because they are getting the JAXBElement form the
reader and not a file. 

I have added a property to the applications.properties file. 

spring.batch.job.names=batch

This is for my verison (process the folder and then shut down) 

spring.batch.job.names=service

is for your version, i.e. it runs until you kill it.

If you don’t want to modifiy the properties file you can add it as a
argument:

java -jar -Dspring.batch.job.names=batch Yoink-0.0.1.jar

that will override what is in the application.properties file

Let me know what you think….